requirements.txt
beautifulsoup4==4.12.2
nltk==3.8.1
requests==2.31.0
scikit-learn==1.3.0
streamlit==1.27.0
PyPDF2==3.0.0
python-docx==0.8.11
